Final fixes
