+++ apps/web/src/app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,5 @@
-import { withPosthog } from "@rallly/posthog/server";
-
 import { handlers } from "@/next-auth";
+import { withPosthog } from "@/utils/posthog";
 
 export const GET = withPosthog(handlers.GET);
 export const POST = withPosthog(handlers.POST);
