+++ packages/emails/package.json
@@ -22,7 +22,7 @@
     "i18next-icu": "^2.3.0",
     "i18next-resources-to-backend": "^1.2.1",
     "nodemailer": "^6.9.9",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-email": "^4.0.2",
     "react-i18next": "^15.5.1"
   },
