+++ packages/posthog/src/server/index.ts
@@ -1,5 +1,3 @@
-import { waitUntil } from "@vercel/functions";
-import type { NextRequest } from "next/server";
 import { PostHog } from "posthog-node";
 
 function PostHogClient() {
@@ -14,15 +12,3 @@ function PostHogClient() {
 }
 
 export const posthog = PostHogClient();
-
-export function withPosthog(handler: (req: NextRequest) => Promise<Response>) {
-  return async (req: NextRequest) => {
-    const res = await handler(req);
-    try {
-      waitUntil(Promise.all([posthog?.shutdown()]));
-    } catch (error) {
-      console.error("Failed to flush PostHog events:", error);
-    }
-    return res;
-  };
-}
